+++ sregistry/database/__init__.py
@@ -5,9 +5,11 @@ if SREGISTRY_DATABASE is None:
 else:
     from .models import *
     from .sqlite import ( 
-        add, get, rm, rmi, images, 
+        add, cp, get, mv, rm, rmi, 
+        images, 
         inspect,
         get_container,
         get_collection,
-        get_or_create_collection 
+        get_or_create_collection,
+        rename
     )