+++ sregistry/main/docker/pull.py
@@ -23,6 +23,7 @@ from sregistry.logger import bot
 from sregistry.client import Singularity
 from sregistry.utils import ( parse_image_name, remove_uri, extract_tar )
 import tempfile
+import shutil
 import os
 import sys
 
@@ -65,15 +66,6 @@ def pull(self, images, file_name=None, save=True, force=False, **kwargs):
 
         digest = q['version'] or q['tag']
 
-        # This is the Docker Hub namespace and repository
-        layers = self._download_layers(q['url'], digest)
-
-        # This is the url where the manifests were obtained
-        url = self._get_manifest_selfLink(q['url'], digest)
-
-        # Create client to build from sandbox
-        cli = Singularity()
-
         # Use Singularity to build the image, based on user preference
         if file_name is None:
             file_name = self._get_storage_name(q)
@@ -82,6 +74,15 @@ def pull(self, images, file_name=None, save=True, force=False, **kwargs):
         if os.path.exists(file_name) and force is False:
             bot.error('Image exists! Remove first, or use --force to overwrite')
             sys.exit(1)
+
+        # This is the url where the manifests are obtained
+        url = self._get_manifest_selfLink(q['url'], digest)
+
+        # This is the Docker Hub namespace and repository
+        layers = self._download_layers(q['url'], digest)
+
+        # Create client to build from sandbox
+        cli = Singularity()
 
         # Build from sandbox
         sandbox = tempfile.mkdtemp()
@@ -106,13 +107,17 @@ def pull(self, images, file_name=None, save=True, force=False, **kwargs):
         # Save to local storage
         if save is True:
 
-            container = self.add(image_path = image_file,
-                                 image_name = q['uri'],
-                                 metadata = self.manifests,
-                                 url = url)
+            container = self.add(image_path=image_file,
+                                 image_uri=q['uri'],
+                                 metadata=self.manifests,
+                                 url=url)
 
             # When the container is created, this is the path to the image
             image_file = container.image
+
+        # If the image_file is different from sandbox, remove sandbox
+        if image_file != sandbox:
+            shutil.rmtree(sandbox)
 
         if os.path.exists(image_file):
             bot.debug('Retrieved image file %s' %image_file)