+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and changes prior to that are (unfortunately) done retrospectively. Critical ite
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/sregistry-cli/tree/master) (0.0.x)
+ - adding move (mv) command so client can update database with new path (0.0.67)
+ - adding rename command, which renames image in storage (and w/o path assumes name there)
+ - fixing bug with Docker and Nvidia pull to clean up temporary sandbox build folders
+ - if fromline is not defined in Singularity Recipe returns empty string (for registry pull)
+ - fixing bug with client not being maintained in shell (was doing import twice)
  - fixing bug with Singularity Hub search (0.0.66)
  - added Dropbox backend with push, pull, record, search, share (0.0.65)
  - added ability for shell to take backend as argument (e.g., `sregistry shell dropbox`)